requirements.txt
# Core data processing
pandas>=2.1.0
polars>=0.19.0
pyarrow>=13.0.0
numpy>=1.24.0
scipy>=1.11.0

# ML and modeling
scikit-learn>=1.3.0
xgboost>=2.0.0
sentence-transformers>=2.2.0
umap-learn>=0.5.4
hdbscan>=0.8.33
networkx>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web framework
streamlit>=1.28.0
openpyxl>=3.1.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
pydantic>=2.4.0
pydantic-settings>=2.0.0
joblib>=1.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Excel processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.6.0

# Optional GPU support (comment out if not needed)
# torch>=2.0.0
# transformers>=4.30.0
